Change the way the generic options are parsed.
As stated in `man bash`, during Patern Substitutions in Parameter Expansion,
the pattern is expanded as in a Pathname Expansion. This means that "all
unquoted occurrences of the characters \, ', and " that did not result from one
of the above expansions are removed.".

In our case, this resulted in errors when using something like
`--directory="test dir"`, which, after pattern substitution, was expanded to
`directory=test dir`.

The solution using sed may not be the best one, but it is the simplest I could
think of right now, so it stays as a suggestion, at least.